+++ src/core/hle/kernel/kernel.h
@@ -88,7 +88,7 @@ private:
     u32 CreateNewObjectID();
 
     /// Creates a new process ID, incrementing the internal process ID counter;
-    u32 CreateNewProcessID();
+    u64 CreateNewProcessID();
 
     /// Creates a new thread ID, incrementing the internal thread ID counter.
     u32 CreateNewThreadID();
